# docs/swagger.yaml
basePath: /
definitions:
...
  model.SmartServiceInstance:
    properties:
      created_at:
        description: unix timestamp, set by service on creation
        type: integer
      deleting:
        type: boolean
      description:
        type: string
      design_id:
        type: string
      error:
        description: is set if module-worker notifies the repository about a error
        type: string
      id:
        type: string
      name:
        type: string
      new_release_id:
        type: string
      parameters:
        items:
          $ref: '#/definitions/model.SmartServiceParameter'
        type: array
      ready:
        type: boolean
      release_id:
        type: string
      running_maintenance_ids:
        items:
          type: string
        type: array
      updated_at:
        description: unix timestamp, set by service on creation
        type: integer
      user_id:
        type: string
    type: object
  model.SmartServiceInstanceInfo:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
  model.SmartServiceParameter:
    properties:
      id:
        type: string
      label:
        type: string
      value: {}
      value_label:
        type: string
    type: object
...
info:
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Smart-Service-Repository API
  version: "0.1"
paths:
...
  /instances:
    get:
      description: returns a list of smart-service instances
      parameters:
      - description: limits size of result; 0 means unlimited
        in: query
        name: limit
        type: integer
      - description: offset to be used in combination with limit
        in: query
        name: offset
        type: integer
      - description: describes the sorting in the form of name.asc
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Total-Count:
              description: count of all matching elements; used for pagination
              type: integer
          schema:
            items:
              $ref: '#/definitions/model.SmartServiceInstance'
            type: array
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: returns a list of smart-service instances
      tags:
      - instances
  /instances-by-process-id/{id}:
    get:
      description: get smart-service instance by process-instance-id
      parameters:
      - description: Process-Instance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SmartServiceInstance'
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: get smart-service instance by process-instance-id
      tags:
      - instances
      - process-id
  /instances-by-process-id/{id}/error:
    put:
      consumes:
      - application/json
      description: sets smart-service instance error
      parameters:
      - description: Process-Instance ID
        in: path
        name: id
        required: true
        type: string
      - description: error message (json encoded)
        in: body
        name: message
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: sets smart-service instance error
      tags:
      - instances
      - process-id
      - error
...
  /instances-by-process-id/{id}/user-id:
    get:
      description: get smart-service instance user-id
      parameters:
      - description: Process-Instance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: get smart-service instance user-id
      tags:
      - instances
      - process-id
      - user-id
...
  /instances/{id}:
    delete:
      description: removes a smart-service instance with all modules
      parameters:
      - description: Instance ID
        in: path
        name: id
        required: true
        type: string
      - description: used if module delete information are invalid
        in: query
        name: ignore_module_delete_errors
        type: boolean
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: removes a smart-service instance with all modules
      tags:
      - instances
    get:
      description: returns a smart-service instance
      parameters:
      - description: Instance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SmartServiceInstance'
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: returns a smart-service instance
      tags:
      - instances
  /instances/{id}/error:
    put:
      consumes:
      - application/json
      description: sets smart-service instance error
      parameters:
      - description: Instance ID
        in: path
        name: id
        required: true
        type: string
      - description: error message (json encoded)
        in: body
        name: message
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: sets smart-service instance error
      tags:
      - instances
      - error
  /instances/{id}/info:
    put:
      consumes:
      - application/json
      description: updates smart-service instance parameter
      parameters:
      - description: Instance ID
        in: path
        name: id
        required: true
        type: string
      - description: SmartServiceParameter
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/model.SmartServiceInstanceInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SmartServiceInstance'
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: updates smart-service instance parameter
      tags:
      - instances
      - parameter
...
  /instances/{id}/parameters:
    put:
      consumes:
      - application/json
      description: updates smart-service instance parameter
      parameters:
      - description: sets new release id if set
        in: query
        name: release_id
        type: string
      - description: Instance ID
        in: path
        name: id
        required: true
        type: string
      - description: SmartServiceParameter
        in: body
        name: message
        required: true
        schema:
          items:
            $ref: '#/definitions/model.SmartServiceParameter'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SmartServiceInstance'
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: updates smart-service instance parameter
      tags:
      - instances
      - parameter
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"